Refactor save materialization internals
