+++ project/jimmer-sql/src/main/java/org/babyfish/jimmer/sql/ast/impl/mutation/Saver.java
@@ -6,20 +6,12 @@ import org.babyfish.jimmer.runtime.DraftSpi;
 import org.babyfish.jimmer.runtime.ImmutableSpi;
 import org.babyfish.jimmer.runtime.Internal;
 import org.babyfish.jimmer.sql.InheritanceType;
-import org.babyfish.jimmer.sql.JSqlClient;
 import org.babyfish.jimmer.sql.ast.TypeMatchMode;
-import org.babyfish.jimmer.sql.ast.impl.EntitiesImpl;
 import org.babyfish.jimmer.sql.ast.mutation.*;
-import org.babyfish.jimmer.sql.cache.CacheDisableConfig;
 import org.babyfish.jimmer.sql.fetcher.Fetcher;
-import org.babyfish.jimmer.sql.fetcher.Field;
-import org.babyfish.jimmer.sql.fetcher.impl.FetcherFactory;
-import org.babyfish.jimmer.sql.fetcher.impl.FetcherImpl;
-import org.babyfish.jimmer.sql.fetcher.impl.FetcherImplementor;
 import org.babyfish.jimmer.sql.meta.JoinTemplate;
 import org.babyfish.jimmer.sql.meta.MiddleTable;
 import org.babyfish.jimmer.sql.runtime.JSqlClientImplementor;
-import org.jetbrains.annotations.Nullable;
 
 import java.sql.Connection;
 import java.util.*;
@@ -139,35 +131,6 @@ public class Saver {
         return new BatchSaveResult<>(ctx.affectedRowCountMap, items);
     }
 
-    @SuppressWarnings("unchecked")
-    <E> BatchSaveResult<E> saveAllJoinedMixed(Collection<E> entities) {
-        if (entities.isEmpty()) {
-            return new BatchSaveResult<>(Collections.emptyMap(), Collections.emptyList());
-        }
-        MutationTrigger trigger = ctx.trigger;
-        List<E> newEntities = (List<E>) Internal.produceList(
-                entities,
-                base -> ((ImmutableSpi) base).__type(),
-                drafts -> saveAllJoinedMixedImpl((List<DraftSpi>) drafts),
-                trigger == null ? null : trigger::prepareSubmit
-        );
-        if (trigger != null) {
-            trigger.submit(ctx.options.getSqlClient(), ctx.con);
-        }
-        Iterator<E> oldItr = entities.iterator();
-        Iterator<E> newItr = newEntities.iterator();
-        List<BatchSaveResult.Item<E>> items = new ArrayList<>(entities.size());
-        while (oldItr.hasNext() && newItr.hasNext()) {
-            items.add(
-                    new BatchSaveResult.Item<>(
-                            oldItr.next(),
-                            newItr.next()
-                    )
-            );
-        }
-        return new BatchSaveResult<>(ctx.affectedRowCountMap, items);
-    }
-
     private void saveAllJoinedInsertImpl(List<DraftSpi> drafts) {
         if (drafts.isEmpty()) {
             return;
@@ -199,159 +162,6 @@ public class Saver {
         for (int i = 0; i < operations.size(); i++) {
             SaveOperation operation = operations.get(i);
             operation.saver.finishSave(operation, selfResults.get(i));
-        }
-    }
-
-    private void saveAllJoinedMixedImpl(List<DraftSpi> drafts) {
-        if (drafts.isEmpty()) {
-            return;
-        }
-        ImmutableType rootType = ctx.path.getType();
-        InheritanceInfo inheritanceInfo = rootType.getInheritanceInfo();
-        if (inheritanceInfo == null || inheritanceInfo.getStrategy() != InheritanceType.JOINED) {
-            throw new AssertionError("Internal bug: mixed joined save must target a joined inheritance root");
-        }
-        Map<ImmutableType, List<DraftSpi>> groupMap = new LinkedHashMap<>();
-        for (DraftSpi draft : drafts) {
-            validateInstantiableSaveType(draft.__type(), ctx.options);
-            groupMap.computeIfAbsent(draft.__type(), it -> new ArrayList<>()).add(draft);
-        }
-        List<SaveOperation> operations = new ArrayList<>(groupMap.size());
-        for (Map.Entry<ImmutableType, List<DraftSpi>> e : groupMap.entrySet()) {
-            SaveContext groupCtx = new SaveContext(
-                    ctx.options,
-                    ctx.con,
-                    e.getKey(),
-                    ctx.fetcher,
-                    ctx.trigger,
-                    ctx.affectedRowCountMap
-            );
-            Saver groupSaver = new Saver(groupCtx);
-            operations.add(groupSaver.prepareSave(e.getValue()));
-        }
-        List<SaveSelfResult> selfResults = saveJoinedMixedSelf(rootType, inheritanceInfo, operations);
-        for (int i = 0; i < operations.size(); i++) {
-            SaveOperation operation = operations.get(i);
-            operation.saver.finishAssociations(operation, selfResults.get(i));
-        }
-        Set<DraftSpi> acceptedDrafts = acceptedDrafts(operations, selfResults);
-        fetch(
-                fetchDrafts(drafts, acceptedDrafts),
-                batches(operations, selfResults)
-        );
-    }
-
-    private List<SaveSelfResult> saveJoinedMixedSelf(
-            ImmutableType rootType,
-            InheritanceInfo inheritanceInfo,
-            List<SaveOperation> operations
-    ) {
-        Map<RootStageKey, RootStageGroup> rootGroupMap = new LinkedHashMap<>();
-        boolean detach = false;
-        for (SaveOperation operation : operations) {
-            for (Batch<DraftSpi> batch : operation.preHandler.batches()) {
-                if (isDetachRequired(batch.mode())) {
-                    detach = true;
-                }
-                DraftSpi sample = batch.entities().iterator().next();
-                Shape rootShape = Operator.joinedRootShape(ctx.options.getSqlClient(), rootType, sample);
-                RootStageKey key = new RootStageKey(rootShape, batch.mode(), batch.originalMode());
-                rootGroupMap.computeIfAbsent(key, it -> new RootStageGroup(rootShape, batch.mode(), batch.originalMode()))
-                        .entities
-                        .addAll(batch.entities());
-            }
-        }
-        Set<DraftSpi> acceptedDrafts = Collections.newSetFromMap(new IdentityHashMap<>());
-        Operator rootOperator = new Operator(ctx);
-        for (RootStageGroup group : rootGroupMap.values()) {
-            int[] rowCounts = saveJoinedRootStage(rootOperator, group, inheritanceInfo);
-            collectAcceptedDrafts(acceptedDrafts, group.entities, rowCounts);
-        }
-        saveJoinedChildStages(rootType, inheritanceInfo, operations, acceptedDrafts);
-        List<SaveSelfResult> results = new ArrayList<>(operations.size());
-        for (SaveOperation operation : operations) {
-            Set<DraftSpi> operationAcceptedDrafts = Collections.newSetFromMap(new IdentityHashMap<>());
-            for (Batch<DraftSpi> batch : operation.preHandler.batches()) {
-                for (DraftSpi draft : batch.entities()) {
-                    if (acceptedDrafts.contains(draft)) {
-                        operationAcceptedDrafts.add(draft);
-                    }
-                }
-            }
-            results.add(new SaveSelfResult(detach, operationAcceptedDrafts));
-        }
-        return results;
-    }
-
-    private int[] saveJoinedRootStage(
-            Operator rootOperator,
-            RootStageGroup group,
-            InheritanceInfo inheritanceInfo
-    ) {
-        Batch<DraftSpi> batch = batchOf(group.shape, group.entities, group.mode, group.originalMode);
-        switch (group.mode) {
-            case INSERT_ONLY:
-                rootOperator.insertJoinedRoot(batch, inheritanceInfo);
-                return acceptedRowCounts(group.entities.size());
-            case INSERT_IF_ABSENT:
-                return rootOperator.upsertJoinedRootStage(batch, inheritanceInfo, true, false, false);
-            case UPDATE_ONLY:
-                return rootOperator.updateJoinedRootStage(null, null, batch, inheritanceInfo, false, false);
-            default:
-                return rootOperator.upsertJoinedRootStage(batch, inheritanceInfo, false, true, false);
-        }
-    }
-
-    private void saveJoinedChildStages(
-            ImmutableType rootType,
-            InheritanceInfo inheritanceInfo,
-            List<SaveOperation> operations,
-            Set<DraftSpi> acceptedDrafts
-    ) {
-        JSqlClientImplementor sqlClient = ctx.options.getSqlClient();
-        Map<ChildStageKey, ChildStageGroup> groupMap = new LinkedHashMap<>();
-        for (SaveOperation operation : operations) {
-            for (Batch<DraftSpi> batch : operation.preHandler.batches()) {
-                for (DraftSpi draft : batch.entities()) {
-                    if (!acceptedDrafts.contains(draft)) {
-                        continue;
-                    }
-                    ImmutableType previousTableType = rootType;
-                    for (ImmutableType tableType : Operator.joinedTableTypes(rootType, draft.__type())) {
-                        Shape shape = Operator.joinedStageShape(sqlClient, previousTableType, tableType, draft);
-                        ChildStageKey key = new ChildStageKey(tableType, shape, batch.mode(), batch.originalMode());
-                        groupMap.computeIfAbsent(
-                                key,
-                                it -> new ChildStageGroup(tableType, shape, batch.mode(), batch.originalMode())
-                        ).entities.add(draft);
-                        previousTableType = tableType;
-                    }
-                }
-            }
-        }
-        for (ChildStageGroup group : groupMap.values()) {
-            SaveContext stageCtx = new SaveContext(
-                    ctx.options,
-                    ctx.con,
-                    group.tableType,
-                    ctx.fetcher,
-                    ctx.trigger,
-                    ctx.affectedRowCountMap
-            );
-            Operator operator = new Operator(stageCtx);
-            Batch<DraftSpi> batch = batchOf(group.shape, group.entities, group.mode, group.originalMode);
-            switch (group.mode) {
-                case INSERT_ONLY:
-                case INSERT_IF_ABSENT:
-                    operator.insertJoinedStage(batch, group.tableType);
-                    break;
-                case UPDATE_ONLY:
-                    operator.updateJoinedStage(batch, group.tableType, rootType, inheritanceInfo);
-                    break;
-                default:
-                    operator.upsertJoinedStage(batch, group.tableType);
-                    break;
-            }
         }
     }
 
@@ -411,7 +221,7 @@ public class Saver {
         finishSave(operation, selfResult);
     }
 
-    private SaveOperation prepareSave(List<DraftSpi> drafts) {
+    SaveOperation prepareSave(List<DraftSpi> drafts) {
         if (!drafts.isEmpty() && !isIdOnlyAssociationReference(drafts)) {
             validateInstantiableSaveType(drafts.get(0).__type(), ctx.options);
         }
@@ -432,16 +242,16 @@ public class Saver {
 
     private void finishSave(SaveOperation operation, SaveSelfResult selfResult) {
         finishAssociations(operation, selfResult);
-        fetch(
-                fetchDrafts(operation.drafts, selfResult.acceptedDrafts),
-                batches(operation.preHandler, selfResult.acceptedDrafts)
+        new SaveResultMaterializer(ctx).materialize(
+                SaveBatches.drafts(operation.drafts, selfResult.acceptedDrafts),
+                SaveBatches.selfBatches(operation.preHandler, selfResult.acceptedDrafts)
         );
     }
 
-    private void finishAssociations(SaveOperation operation, SaveSelfResult selfResult) {
+    void finishAssociations(SaveOperation operation, SaveSelfResult selfResult) {
         PreHandler preHandler = operation.preHandler;
 
-        for (Batch<DraftSpi> batch : associationBatches(preHandler, selfResult.acceptedDrafts)) {
+        for (Batch<DraftSpi> batch : SaveBatches.associationBatches(preHandler, selfResult.acceptedDrafts)) {
             for (ImmutableProp prop : batch.shape().getGetterMap().keySet()) {
                 if (isVisitable(prop) && prop.isAssociation(TargetLevel.ENTITY)) {
                     if (ctx.options.getAssociatedMode(prop) == AssociatedSaveMode.VIOLENTLY_REPLACE) {
@@ -552,376 +362,6 @@ public class Saver {
         updateAssociations(batch, prop, detachOtherSiblings);
     }
 
-    private void fetch(List<DraftSpi> drafts, Iterable<Batch<DraftSpi>> batches) {
-        if (ctx.path.getParent() != null) {
-            fetchIdIfNecessary(drafts, batches);
-            return;
-        }
-        Fetcher<?> fetcher = ctx.fetcher;
-        if (fetcher == null) {
-            fetchIdIfNecessary(drafts, batches);
-            return;
-        }
-        fetchImpl(drafts, batches, false);
-    }
-
-    private void fetchIdIfNecessary(List<DraftSpi> drafts, Iterable<Batch<DraftSpi>> batches) {
-        if (ctx.options.getMode() != SaveMode.INSERT_IF_ABSENT) {
-            return;
-        }
-        if (!ctx.isIdRetrievingRequired()) {
-            return;
-        }
-        fetchImpl(drafts, batches, true);
-    }
-
-    @SuppressWarnings("unchecked")
-    private void fetchImpl(
-            List<DraftSpi> drafts,
-            Iterable<Batch<DraftSpi>> batches,
-            boolean fillIdIfNecessary
-    ) {
-        DraftSpi[] arr = new DraftSpi[drafts.size()];
-        int index = 0;
-        List<Object> unmatchedIds = new ArrayList<>();
-        List<DraftSpi> nonIdObjects = new ArrayList<>();
-        Fetcher<?> fetcher = ctx.fetcher;
-        ResidualFetchGroup residualGroup = null;
-        DraftSpi[] databaseDefaultArr = new DraftSpi[drafts.size()];
-        List<Object> databaseDefaultIds = new ArrayList<>();
-        SaveFetcherAnalysis fetcherAnalysis = fetcher != null ?
-                SaveFetcherAnalysis.of(fetcher) :
-                null;
-        List<ImmutableProp> databaseDefaultProps =
-                fetcherAnalysis != null && fetcherAnalysis.isScalarOnly() ?
-                        fetcherAnalysis.getDatabaseDefaultProps() :
-                        Collections.emptyList();
-        boolean canFetchDatabaseDefaults = fetcherAnalysis != null && !databaseDefaultProps.isEmpty();
-        Fetcher<Object> residualFetcher = fetcher != null &&
-                fetcherAnalysis != null &&
-                !fillIdIfNecessary ?
-                residualFetcher(fetcher) :
-                null;
-        if (residualFetcher != null) {
-            residualGroup = new ResidualFetchGroup(residualFetcher);
-        }
-        PropId idPropId = ctx.path.getType().getIdProp().getId();
-        SaveShapeMatcher shapeMatcher = new SaveShapeMatcher(ctx.options::getUpsertMask);
-        for (DraftSpi draft : drafts) {
-            if (ctx.isSaveReturningNotAccepted(draft)) {
-                // Returning row-count 0 rows must remain unmaterialized.
-                ++index;
-                continue;
-            } else if (!draft.__isLoaded(idPropId)) {
-                nonIdObjects.add(draft);
-            } else if (fetcher != null && !shapeMatcher.isMatched(draft, fetcher, true)) {
-                Object id = draft.__get(idPropId);
-                if (canFetchDatabaseDefaults &&
-                        fetcherAnalysis.isUnmatchedOnlyByDatabaseDefaultProps(draft)) {
-                    databaseDefaultArr[index] = draft;
-                    databaseDefaultIds.add(id);
-                } else if (residualFetcher != null &&
-                        ctx.isSaveReturningApplied(draft) &&
-                        fetcherAnalysis.areReturningPropsLoaded(draft)) {
-                    residualGroup.add(index, draft, id);
-                } else {
-                    arr[index] = draft;
-                    unmatchedIds.add(id);
-                }
-            }
-            ++index;
-        }
-        if (!databaseDefaultIds.isEmpty()) {
-            JSqlClient sqlClient = ctx.options.getSqlClient().caches(CacheDisableConfig::disableAll);
-            Fetcher<ImmutableSpi> databaseDefaultFetcher =
-                    databaseDefaultFetcher(databaseDefaultProps);
-            Map<Object, ImmutableSpi> map = ((EntitiesImpl) sqlClient.getEntities())
-                    .forSaveCommandFetch(QueryReason.FETCHER)
-                    .forConnection(ctx.con)
-                    .findMapByIds(databaseDefaultFetcher, databaseDefaultIds);
-            index = 0;
-            for (DraftSpi draft : databaseDefaultArr) {
-                if (draft != null) {
-                    Object id = draft.__get(idPropId);
-                    ImmutableSpi fetched = map.get(id);
-                    if (fetched != null) {
-                        applyDatabaseDefaultProps(draft, fetched, databaseDefaultProps);
-                    } else {
-                        arr[index] = draft;
-                        unmatchedIds.add(id);
-                    }
-                }
-                ++index;
-            }
-        }
-        if (residualGroup != null && !residualGroup.ids.isEmpty()) {
-            fetchResidual(residualGroup, arr, unmatchedIds, shapeMatcher, fetcher, idPropId);
-        }
-        if (!unmatchedIds.isEmpty()) {
-            JSqlClient sqlClient = ctx.options.getSqlClient().caches(CacheDisableConfig::disableAll);
-            Map<Object, Object> map = ((EntitiesImpl) sqlClient.getEntities())
-                    .forSaveCommandFetch(fillIdIfNecessary ? QueryReason.GET_ID_FOR_PRE_SAVED_ENTITIES : QueryReason.FETCHER)
-                    .forConnection(ctx.con)
-                    .findMapByIds(
-                            fillIdIfNecessary ?
-                                    new FetcherImpl<>((Class<Object>) ctx.path.getType().getJavaClass()) :
-                                    (Fetcher<Object>) fetcher,
-                            unmatchedIds
-                    );
-            index = 0;
-            ListIterator<DraftSpi> itr = drafts.listIterator();
-            while (itr.hasNext()) {
-                DraftSpi draft = itr.next();
-                if (arr[index] != null) {
-                    Object fetched = map.get(draft.__get(idPropId));
-                    DraftSpi replacedDraft = replaceDraft(draft, fetched);
-                    if (replacedDraft != null) {
-                        itr.set(replacedDraft);
-                    }
-                }
-                ++index;
-            }
-        }
-        if (!nonIdObjects.isEmpty()) {
-            if (drafts.size() == 1) {
-                Map<KeyMatcher.Group, List<ImmutableSpi>> rowMap = Rows.findByKeys(
-                        ctx,
-                        fillIdIfNecessary ? QueryReason.GET_ID_FOR_PRE_SAVED_ENTITIES : QueryReason.FETCHER,
-                        fillIdIfNecessary ?
-                                new FetcherImpl<>((Class<ImmutableSpi>) ctx.path.getType().getJavaClass()) :
-                                (Fetcher<ImmutableSpi>) fetcher,
-                        nonIdObjects,
-                        null
-                );
-                if (!rowMap.isEmpty()) {
-                    ImmutableSpi row = rowMap.values().iterator().next().iterator().next();
-                    if (fillIdIfNecessary) {
-                        drafts.get(0).__set(idPropId, row.__get(idPropId));
-                    } else {
-                        ListIterator<DraftSpi> itr = drafts.listIterator();
-                        DraftSpi draft = itr.next();
-                        DraftSpi replaceDraft = replaceDraft(draft, row);
-                        if (replaceDraft != null) {
-                            itr.set(replaceDraft);
-                        }
-                    }
-                }
-            } else {
-                KeyMatcher keyMatcher = ctx.options.getKeyMatcher(ctx.path.getType());
-                for (Batch<DraftSpi> batch : batches) {
-                    Set<ImmutableProp> keyProps = batch.shape().keyProps(keyMatcher);
-                    List<PropId> unloadPropIds = null;
-                    if (!fillIdIfNecessary) {
-                        assert fetcher != null;
-                        unloadPropIds = new ArrayList<>();
-                        for (ImmutableProp keyProp : keyProps) {
-                            if (!((FetcherImplementor<?>) fetcher).__contains(keyProp.getName())) {
-                                fetcher = fetcher.add(keyProp.getName());
-                                unloadPropIds.add(keyProp.getId());
-                            }
-                        }
-                    }
-                    Fetcher<ImmutableSpi> actualFetcher;
-                    if (fillIdIfNecessary) {
-                        actualFetcher = new FetcherImpl<>((Class<ImmutableSpi>) ctx.path.getType().getJavaClass());
-                        for (ImmutableProp keyProp : keyProps) {
-                            actualFetcher = actualFetcher.add(keyProp.getName());
-                        }
-                    } else {
-                        actualFetcher = (Fetcher<ImmutableSpi>) fetcher;
-                    }
-                    Map<KeyMatcher.Group, Map<Object, ImmutableSpi>> map = Rows.findMapByKeys(
-                            ctx,
-                            fillIdIfNecessary ? QueryReason.GET_ID_FOR_PRE_SAVED_ENTITIES : QueryReason.FETCHER,
-                            actualFetcher,
-                            nonIdObjects
-                    );
-                    if (map.isEmpty()) {
-                        continue;
-                    }
-                    ListIterator<DraftSpi> itr = drafts.listIterator();
-                    while (itr.hasNext()) {
-                        DraftSpi draft = itr.next();
-                        if (draft.__isLoaded(idPropId)) {
-                            continue;
-                        }
-                        Map<Object, ImmutableSpi> subMap = map.values().iterator().next();
-                        Object key = Keys.keyOf(draft, keyProps);
-                        ImmutableSpi fetched = subMap.get(key);
-                        if (unloadPropIds == null) {
-                            draft.__set(idPropId, fetched.__get(idPropId));
-                        } else {
-                            DraftSpi newDraft = replaceDraft(draft, fetched);
-                            if (newDraft != null) {
-                                itr.set(newDraft);
-                            } else {
-                                newDraft = draft;
-                            }
-                            for (PropId unloadedPropId : unloadPropIds) {
-                                newDraft.__unload(unloadedPropId);
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    @SuppressWarnings("unchecked")
-    private void fetchResidual(
-            ResidualFetchGroup group,
-            DraftSpi[] arr,
-            List<Object> unmatchedIds,
-            SaveShapeMatcher shapeMatcher,
-            Fetcher<?> fetcher,
-            PropId idPropId
-    ) {
-        JSqlClient sqlClient = ctx.options.getSqlClient().caches(CacheDisableConfig::disableAll);
-        Fetcher<Object> residualFetcher = residualFetcher(group.fetcher, group.types);
-        if (residualFetcher == null) {
-            return;
-        }
-        Map<Object, Object> map = ((EntitiesImpl) sqlClient.getEntities())
-                .forSaveCommandFetch(QueryReason.FETCHER)
-                .forConnection(ctx.con)
-                .findMapByIds(residualFetcher, group.ids);
-        for (int i = 0; i < group.drafts.size(); i++) {
-            DraftSpi draft = group.drafts.get(i);
-            Object id = draft.__get(idPropId);
-            Object fetched = map.get(id);
-            if (mergeDraft(draft, fetched) &&
-                    shapeMatcher.isMatched(draft, fetcher, false)) {
-                shapeMatcher.trim(draft, fetcher);
-            } else {
-                arr[group.indexes.get(i)] = draft;
-                unmatchedIds.add(id);
-            }
-        }
-    }
-
-    @SuppressWarnings("unchecked")
-    private static Fetcher<Object> residualFetcher(Fetcher<?> fetcher) {
-        return residualFetcher(fetcher, Collections.emptySet());
-    }
-
-    @SuppressWarnings("unchecked")
-    private static Fetcher<Object> residualFetcher(Fetcher<?> fetcher, Collection<ImmutableType> types) {
-        boolean hasTypeBranches = !((FetcherImplementor<?>) fetcher).__getTypeBranchFetcherMap().isEmpty();
-        ImmutableType rootType = fetcher.getImmutableType();
-        FetcherFactory.PropFilter propFilter = hasTypeBranches ?
-                Saver::isPolymorphicResidualField :
-                (type, prop, path) -> !path.isEmpty() || !SaveFetcherAnalysis.isScalarColumnProp(prop);
-        Fetcher<Object> residualFetcher = (Fetcher<Object>) FetcherFactory.filterByTypedProp(
-                (Fetcher<Object>) fetcher,
-                hasTypeBranches && !types.isEmpty() ?
-                        (type, path) -> type == rootType || containsAssignableType(types, type) :
-                        null,
-                propFilter
-        );
-        return isIdOnlyFetcher(residualFetcher) ? null : residualFetcher;
-    }
-
-    private static boolean containsAssignableType(Collection<ImmutableType> types, ImmutableType type) {
-        for (ImmutableType actualType : types) {
-            if (type.isAssignableFrom(actualType)) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    private static boolean isPolymorphicResidualField(
-            ImmutableType type,
-            ImmutableProp prop,
-            List<ImmutableProp> path
-    ) {
-        if (!path.isEmpty() || !SaveFetcherAnalysis.isScalarColumnProp(prop)) {
-            return true;
-        }
-        InheritanceInfo inheritanceInfo = type.getInheritanceInfo();
-        return inheritanceInfo != null && !inheritanceInfo.isPropAvailableInTable(prop, inheritanceInfo.getRootType());
-    }
-
-    private static boolean isIdOnlyFetcher(Fetcher<?> fetcher) {
-        if (!((FetcherImplementor<?>) fetcher).__getTypeBranchFetcherMap().isEmpty()) {
-            return false;
-        }
-        Map<String, Field> fieldMap = fetcher.getFieldMap();
-        return fieldMap.size() == 1 && fieldMap.values().iterator().next().getProp().isId();
-    }
-
-    private Fetcher<ImmutableSpi> databaseDefaultFetcher(List<ImmutableProp> props) {
-        Fetcher<ImmutableSpi> fetcher = new FetcherImpl<>((Class<ImmutableSpi>) ctx.path.getType().getJavaClass());
-        for (ImmutableProp prop : props) {
-            fetcher = fetcher.add(prop.getName());
-        }
-        return fetcher;
-    }
-
-    private static void applyDatabaseDefaultProps(
-            DraftSpi draft,
-            ImmutableSpi fetched,
-            List<ImmutableProp> props
-    ) {
-        for (ImmutableProp prop : props) {
-            PropId propId = prop.getId();
-            if (fetched.__isLoaded(propId)) {
-                draft.__set(propId, fetched.__get(propId));
-                draft.__show(propId, fetched.__isVisible(propId));
-            }
-        }
-    }
-
-    private static DraftSpi replaceDraft(DraftSpi draft, Object fetched) {
-        if (fetched instanceof DraftSpi) {
-            return (DraftSpi) fetched;
-        }
-        if (fetched instanceof ImmutableSpi) {
-            ImmutableSpi spi = (ImmutableSpi) fetched;
-            for (ImmutableProp prop : draft.__type().getProps().values()) {
-                PropId propId = prop.getId();
-                if (spi.__isLoaded(propId)) {
-                    if (!prop.isView()) {
-                        draft.__set(propId, spi.__get(propId));
-                    }
-                    draft.__show(propId, spi.__isVisible(propId));
-                } else {
-                    draft.__unload(propId);
-                }
-            }
-        }
-        return null;
-    }
-
-    private static boolean mergeDraft(DraftSpi draft, Object fetched) {
-        if (fetched instanceof ImmutableSpi) {
-            ImmutableSpi spi = (ImmutableSpi) fetched;
-            for (ImmutableProp prop : draft.__type().getProps().values()) {
-                PropId propId = prop.getId();
-                if (spi.__isLoaded(propId)) {
-                    if (!prop.isView()) {
-                        Object value = spi.__get(propId);
-                        if (value != null) {
-                            if (prop.isReferenceList(TargetLevel.OBJECT)) {
-                                value = draft.__draftContext().toDraftList(
-                                        (List<Object>) value,
-                                        (Class<Object>) prop.getElementClass(),
-                                        true
-                                );
-                            } else if (prop.isReference(TargetLevel.OBJECT)) {
-                                value = draft.__draftContext().toDraftObject(value);
-                            }
-                        }
-                        draft.__set(propId, value);
-                    }
-                    draft.__show(propId, spi.__isVisible(propId));
-                }
-            }
-            return true;
-        }
-        return false;
-    }
-
     private boolean isVisitable(ImmutableProp prop) {
         ImmutableProp backRef = ctx.backReferenceProp;
         return backRef == null || backRef != prop;
@@ -981,7 +421,7 @@ public class Saver {
         Operator operator = new Operator(ctx);
         for (StageGroup group : groupMap.values()) {
             operator.insertJoinedRoot(
-                    batchOf(group.shape, group.entities, SaveMode.INSERT_ONLY, SaveMode.INSERT_ONLY),
+                    SaveBatches.of(group.shape, group.entities, SaveMode.INSERT_ONLY, SaveMode.INSERT_ONLY),
                     inheritanceInfo
             );
         }
@@ -1016,58 +456,9 @@ public class Saver {
                     ctx.affectedRowCountMap
             );
             new Operator(stageCtx).insertJoinedStage(
-                    batchOf(group.shape, group.entities, SaveMode.INSERT_ONLY, SaveMode.INSERT_ONLY),
+                    SaveBatches.of(group.shape, group.entities, SaveMode.INSERT_ONLY, SaveMode.INSERT_ONLY),
                     group.tableType
             );
-        }
-    }
-
-    private static Batch<DraftSpi> batchOf(
-            Shape shape,
-            EntityCollection<DraftSpi> entities,
-            SaveMode mode,
-            SaveMode originalMode
-    ) {
-        return new Batch<DraftSpi>() {
-
-            @Override
-            public Shape shape() {
-                return shape;
-            }
-
-            @Override
-            public EntityCollection<DraftSpi> entities() {
-                return entities;
-            }
-
-            @Override
-            public SaveMode mode() {
-                return mode;
-            }
-
-            @Override
-            public SaveMode originalMode() {
-                return originalMode;
-            }
-        };
-    }
-
-    private static int[] acceptedRowCounts(int size) {
-        int[] rowCounts = new int[size];
-        Arrays.fill(rowCounts, 1);
-        return rowCounts;
-    }
-
-    private static void collectAcceptedDrafts(
-            Set<DraftSpi> output,
-            EntityCollection<DraftSpi> entities,
-            int[] rowCounts
-    ) {
-        int index = 0;
-        for (EntityCollection.Item<DraftSpi> item : entities.items()) {
-            if (index < rowCounts.length && rowCounts[index++] != 0) {
-                output.add(item.getEntity());
-            }
         }
     }
 
@@ -1116,7 +507,7 @@ public class Saver {
         }
     }
 
-    private static boolean isDetachRequired(SaveMode mode) {
+    static boolean isDetachRequired(SaveMode mode) {
         return mode != SaveMode.INSERT_ONLY && mode != SaveMode.INSERT_IF_ABSENT;
     }
 
@@ -1168,88 +559,6 @@ public class Saver {
         return inheritanceInfo != null &&
                 inheritanceInfo.getStrategy() == InheritanceType.JOINED &&
                 inheritanceInfo.getRootType() != type;
-    }
-
-    private Iterable<Batch<DraftSpi>> associationBatches(
-            PreHandler preHandler,
-            @Nullable Set<DraftSpi> acceptedDrafts
-    ) {
-        return acceptedDrafts != null ?
-                filterBatches(preHandler.associationBatches(), acceptedDrafts) :
-                preHandler.associationBatches();
-    }
-
-    private Iterable<Batch<DraftSpi>> batches(
-            PreHandler preHandler,
-            @Nullable Set<DraftSpi> acceptedDrafts
-    ) {
-        return acceptedDrafts != null ?
-                filterBatches(preHandler.batches(), acceptedDrafts) :
-                preHandler.batches();
-    }
-
-    private Iterable<Batch<DraftSpi>> batches(
-            List<SaveOperation> operations,
-            List<SaveSelfResult> selfResults
-    ) {
-        List<Batch<DraftSpi>> batches = new ArrayList<>();
-        for (int i = 0; i < operations.size(); i++) {
-            SaveOperation operation = operations.get(i);
-            SaveSelfResult selfResult = selfResults.get(i);
-            for (Batch<DraftSpi> batch : batches(operation.preHandler, selfResult.acceptedDrafts)) {
-                batches.add(batch);
-            }
-        }
-        return batches;
-    }
-
-    private static Set<DraftSpi> acceptedDrafts(
-            List<SaveOperation> operations,
-            List<SaveSelfResult> selfResults
-    ) {
-        Set<DraftSpi> acceptedDrafts = Collections.newSetFromMap(new IdentityHashMap<>());
-        for (int i = 0; i < operations.size(); i++) {
-            SaveOperation operation = operations.get(i);
-            Set<DraftSpi> operationAcceptedDrafts = selfResults.get(i).acceptedDrafts;
-            if (operationAcceptedDrafts != null) {
-                acceptedDrafts.addAll(operationAcceptedDrafts);
-            } else {
-                acceptedDrafts.addAll(operation.drafts);
-            }
-        }
-        return acceptedDrafts;
-    }
-
-    private List<DraftSpi> fetchDrafts(List<DraftSpi> drafts, @Nullable Set<DraftSpi> acceptedDrafts) {
-        if (acceptedDrafts == null) {
-            return drafts;
-        }
-        List<DraftSpi> filteredDrafts = new ArrayList<>(acceptedDrafts.size());
-        for (DraftSpi draft : drafts) {
-            if (acceptedDrafts.contains(draft)) {
-                filteredDrafts.add(draft);
-            }
-        }
-        return filteredDrafts;
-    }
-
-    private Iterable<Batch<DraftSpi>> filterBatches(
-            Iterable<Batch<DraftSpi>> batches,
-            Set<DraftSpi> acceptedDrafts
-    ) {
-        List<Batch<DraftSpi>> filteredBatches = new ArrayList<>();
-        for (Batch<DraftSpi> batch : batches) {
-            EntityList<DraftSpi> entities = new EntityList<>();
-            for (DraftSpi draft : batch.entities()) {
-                if (acceptedDrafts.contains(draft)) {
-                    entities.add(draft);
-                }
-            }
-            if (!entities.isEmpty()) {
-                filteredBatches.add(new FilteredBatch(batch, entities));
-            }
-        }
-        return filteredBatches;
     }
 
     private void clearAssociations(Collection<? extends ImmutableSpi> rows, ImmutableProp prop) {
@@ -1364,182 +673,6 @@ public class Saver {
         return false;
     }
 
-    private static class SaveOperation {
-
-        final Saver saver;
-
-        final List<DraftSpi> drafts;
-
-        final PreHandler preHandler;
-
-        final boolean ownerAcceptanceRequired;
-
-        SaveOperation(
-                Saver saver,
-                List<DraftSpi> drafts,
-                PreHandler preHandler,
-                boolean ownerAcceptanceRequired
-        ) {
-            this.saver = saver;
-            this.drafts = drafts;
-            this.preHandler = preHandler;
-            this.ownerAcceptanceRequired = ownerAcceptanceRequired;
-        }
-    }
-
-    private static class SaveSelfResult {
-
-        final boolean detach;
-
-        @Nullable
-        final Set<DraftSpi> acceptedDrafts;
-
-        SaveSelfResult(boolean detach, @Nullable Set<DraftSpi> acceptedDrafts) {
-            this.detach = detach;
-            this.acceptedDrafts = acceptedDrafts;
-        }
-    }
-
-    private static class RootStageKey {
-
-        final Shape shape;
-
-        final SaveMode mode;
-
-        final SaveMode originalMode;
-
-        RootStageKey(Shape shape, SaveMode mode, SaveMode originalMode) {
-            this.shape = shape;
-            this.mode = mode;
-            this.originalMode = originalMode;
-        }
-
-        @Override
-        public int hashCode() {
-            int hash = shape.hashCode();
-            hash = hash * 31 + mode.hashCode();
-            hash = hash * 31 + originalMode.hashCode();
-            return hash;
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            if (this == obj) {
-                return true;
-            }
-            if (!(obj instanceof RootStageKey)) {
-                return false;
-            }
-            RootStageKey other = (RootStageKey) obj;
-            return shape.equals(other.shape) &&
-                    mode == other.mode &&
-                    originalMode == other.originalMode;
-        }
-    }
-
-    private static class RootStageGroup {
-
-        final Shape shape;
-
-        final SaveMode mode;
-
-        final SaveMode originalMode;
-
-        final EntityList<DraftSpi> entities = new EntityList<>();
-
-        RootStageGroup(Shape shape, SaveMode mode, SaveMode originalMode) {
-            this.shape = shape;
-            this.mode = mode;
-            this.originalMode = originalMode;
-        }
-    }
-
-    private static class ChildStageKey {
-
-        final ImmutableType tableType;
-
-        final Shape shape;
-
-        final SaveMode mode;
-
-        final SaveMode originalMode;
-
-        ChildStageKey(ImmutableType tableType, Shape shape, SaveMode mode, SaveMode originalMode) {
-            this.tableType = tableType;
-            this.shape = shape;
-            this.mode = mode;
-            this.originalMode = originalMode;
-        }
-
-        @Override
-        public int hashCode() {
-            int hash = tableType.hashCode();
-            hash = hash * 31 + shape.hashCode();
-            hash = hash * 31 + mode.hashCode();
-            hash = hash * 31 + originalMode.hashCode();
-            return hash;
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            if (this == obj) {
-                return true;
-            }
-            if (!(obj instanceof ChildStageKey)) {
-                return false;
-            }
-            ChildStageKey other = (ChildStageKey) obj;
-            return tableType == other.tableType &&
-                    shape.equals(other.shape) &&
-                    mode == other.mode &&
-                    originalMode == other.originalMode;
-        }
-    }
-
-    private static class ChildStageGroup {
-
-        final ImmutableType tableType;
-
-        final Shape shape;
-
-        final SaveMode mode;
-
-        final SaveMode originalMode;
-
-        final EntityList<DraftSpi> entities = new EntityList<>();
-
-        ChildStageGroup(ImmutableType tableType, Shape shape, SaveMode mode, SaveMode originalMode) {
-            this.tableType = tableType;
-            this.shape = shape;
-            this.mode = mode;
-            this.originalMode = originalMode;
-        }
-    }
-
-    private static class ResidualFetchGroup {
-
-        final Fetcher<Object> fetcher;
-
-        final List<Integer> indexes = new ArrayList<>();
-
-        final List<DraftSpi> drafts = new ArrayList<>();
-
-        final List<Object> ids = new ArrayList<>();
-
-        final Set<ImmutableType> types = new LinkedHashSet<>();
-
-        ResidualFetchGroup(Fetcher<Object> fetcher) {
-            this.fetcher = fetcher;
-        }
-
-        void add(int index, DraftSpi draft, Object id) {
-            indexes.add(index);
-            drafts.add(draft);
-            ids.add(id);
-            types.add(draft.__type());
-        }
-    }
-
     private static class StageKey {
 
         final ImmutableType tableType;
@@ -1580,38 +713,6 @@ public class Saver {
         StageGroup(ImmutableType tableType, Shape shape) {
             this.tableType = tableType;
             this.shape = shape;
-        }
-    }
-
-    private static class FilteredBatch implements Batch<DraftSpi> {
-
-        private final Batch<DraftSpi> base;
-
-        private final EntityCollection<DraftSpi> entities;
-
-        private FilteredBatch(Batch<DraftSpi> base, EntityCollection<DraftSpi> entities) {
-            this.base = base;
-            this.entities = entities;
-        }
-
-        @Override
-        public Shape shape() {
-            return base.shape();
-        }
-
-        @Override
-        public EntityCollection<DraftSpi> entities() {
-            return entities;
-        }
-
-        @Override
-        public SaveMode mode() {
-            return base.mode();
-        }
-
-        @Override
-        public SaveMode originalMode() {
-            return base.originalMode();
         }
     }
